Move ListDropdownMenu to new utils folder
