+++ app/src/main/java/com/github/geohunt/app/ui/components/challengecreation/ChallengeSettings.kt
@@ -8,6 +8,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.github.geohunt.app.R
 import com.github.geohunt.app.model.database.api.Challenge
+import com.github.geohunt.app.ui.components.utils.ListDropdownMenu
 import com.github.geohunt.app.ui.theme.Lobster
 
 /**
@@ -41,40 +42,6 @@ fun DifficultySelect(selectedDifficulty: MutableState<Challenge.Difficulty>) {
 }
 
 
-/**
- * Creates a dropdown menu containing all the elements of elements.
- * Stores the selected value in the MutableState state to make it accessible to the caller.
- * @param state The mutableState that will be modified by the menu
- * @param elements The elements to chose from
- * @param toString The function used to convert elements to strings that will be displayed
- */
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun <T> ListDropdownMenu(state: MutableState<T>, elements: Collection<T>, toString: (T) -> String) {
-    var expanded by remember { mutableStateOf(false) }
-
-    ExposedDropdownMenuBox(expanded = expanded,
-            onExpandedChange = { expanded = !expanded }) {
-        TextField(value = toString(state.value),
-                onValueChange = {},
-                readOnly = true,
-                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) })
-
-        ExposedDropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false }
-        ) {
-            elements.forEach { elem ->
-                DropdownMenuItem(onClick = {
-                    state.value = elem
-                    expanded = false
-                }) {
-                    Text(text = toString(elem))
-                }
-            }
-        }
-    }
-}
 
 
 @Composable
